+++ pkg/binder/common/env_test.go
@@ -60,3 +60,57 @@ func TestOverrideVar(t *testing.T) {
 	assert.Equal(t, container.Env[0].Name, "myVar")
 	assert.Equal(t, container.Env[0].Value, "myValue1")
 }
+
+func TestOverrideVarWithValueFrom(t *testing.T) {
+	container := &v1.Container{
+		Env: []v1.EnvVar{
+			{
+				Name:  "myVar",
+				Value: "myValue",
+			},
+		},
+	}
+	envVar := v1.EnvVar{
+		Name: "myVar",
+		ValueFrom: &v1.EnvVarSource{
+			ConfigMapKeyRef: &v1.ConfigMapKeySelector{
+				Key: "myVarKeyInCm",
+				LocalObjectReference: v1.LocalObjectReference{
+					Name: "myCMName",
+				},
+			},
+		},
+	}
+	AddEnvVarToContainer(container, envVar)
+	assert.Equal(t, len(container.Env), 1)
+	assert.Equal(t, container.Env[0].Name, "myVar")
+	assert.Equal(t, container.Env[0].Value, "")
+	assert.Equal(t, container.Env[0].ValueFrom, envVar.ValueFrom)
+}
+
+func TestOverrideVarWithValue(t *testing.T) {
+	container := &v1.Container{
+		Env: []v1.EnvVar{
+			{
+				Name: "myVar",
+				ValueFrom: &v1.EnvVarSource{
+					ConfigMapKeyRef: &v1.ConfigMapKeySelector{
+						Key: "myVarKeyInCm",
+						LocalObjectReference: v1.LocalObjectReference{
+							Name: "myCMName",
+						},
+					},
+				},
+			},
+		},
+	}
+	envVar := v1.EnvVar{
+		Name:  "myVar",
+		Value: "myValue",
+	}
+	AddEnvVarToContainer(container, envVar)
+	assert.Equal(t, len(container.Env), 1)
+	assert.Equal(t, container.Env[0].Name, "myVar")
+	assert.Nil(t, container.Env[0].ValueFrom)
+	assert.Equal(t, container.Env[0].Value, envVar.Value)
+}